+++ CHANGELOG.md
@@ -7,16 +7,31 @@ This section contains changes that have been committed but not yet released.
 ### Added
 
 * Added hide participants field for the `Event` class
+* Improved webhook notification support with the addition of more notification attributes and event metadata fields
 
 ### Changed
+
+* NylasAccount.revokeAccessToken() returns a boolean value now
 
 ### Deprecated
 
 ### Fixed
 
+* Fix error when updating occurrence of a recurring event
+
 ### Removed
 
 ### Security
+
+## [1.17.0] - Released 2022-09-22
+
+### Added
+
+* Add support for getting a single expanded message and thread
+
+### Fixed
+
+* Fixed issue where `Event` participants could never be entirely removed once set
 
 ## [1.16.0] - Released 2022-07-29
 
@@ -297,7 +312,8 @@ This second release aims toward API stability so that we can get to v1.0.0.
 
 Initial preview release
 
-[Unreleased]: https://github.com/nylas/nylas-java/compare/v1.16.0...HEAD
+[Unreleased]: https://github.com/nylas/nylas-java/compare/v1.17.0...HEAD
+[1.17.0]: https://github.com/nylas/nylas-java/releases/tag/v1.17.0
 [1.16.0]: https://github.com/nylas/nylas-java/releases/tag/v1.16.0
 [1.15.0]: https://github.com/nylas/nylas-java/releases/tag/v1.15.0
 [1.14.0]: https://github.com/nylas/nylas-java/releases/tag/v1.14.0